# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Conarket",
        "description": "API RESTFul do projeto conarket",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Efetuar o login na API",
                "operationId": "login",
                "requestBody": {
                    "request": "Login",
                    "description": "Credencias do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "LoginResponse",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/password/reset/request": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Requisição de mudança de email",
                "operationId": "sendResetPasswordRequest",
                "requestBody": {
                    "request": "ResetPasswordRequest",
                    "description": "Email para a recuperação de senha",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "ResetPasswordResponse",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Requisição para envia os dados da nova senha",
                "operationId": "resetPassword",
                "requestBody": {
                    "request": "ResetRequest",
                    "description": "Dados da nova senha",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "ResetResponse",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unathorizated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/carts": {
            "get": {
                "tags": [
                    "Carrinhos"
                ],
                "summary": "Retorna os carrinhos do usuário autenticado",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cart"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/carts/{id}": {
            "get": {
                "tags": [
                    "Carrinhos"
                ],
                "summary": "Retorna um carrinho pelo ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Carrinho",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Cart"
                                        },
                                        {
                                            "properties": {
                                                "product_carts": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductCart"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cart": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "total_price": {
                        "description": "Preço total dos items dentro do carrinho de compras",
                        "type": "number"
                    },
                    "closed_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "is_closed": {
                        "type": "boolean"
                    },
                    "user_id": {
                        "description": "Id do usuário dono do carrinho de compras",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "LoginRequest": {
                "properties": {
                    "email": {
                        "description": "Email do usuário",
                        "type": "string"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string"
                    },
                    "user_type": {
                        "description": "ADMIN_COOP | ADMIN_CONAB | CUSTOMER",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ResetPasswordRequest": {
                "properties": {
                    "email": {
                        "description": "Email do usuário",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ResetRequest": {
                "properties": {
                    "email": {
                        "description": "Email do usuário",
                        "type": "string"
                    },
                    "password": {
                        "description": "A nova senha do usuário",
                        "type": "string"
                    },
                    "password_confirmation": {
                        "description": "Confirmação da senha",
                        "type": "string"
                    },
                    "code": {
                        "description": "Email do usuário",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "ProductCart": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "amount": {
                        "description": "Quantidade do produto no carrinho",
                        "type": "number"
                    },
                    "price": {
                        "description": "Preço do produto no carrinho",
                        "type": "number"
                    },
                    "delivered_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "unit_of_measure": {
                        "description": "Unidade de medida do Produto (kg ou unit)",
                        "type": "string"
                    },
                    "product_id": {
                        "description": "Id do produto",
                        "type": "integer"
                    },
                    "cart_id": {
                        "description": "Id do carrinho",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "description": "Token JWT",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer <Token>"
            }
        }
    },
    "security": [
        []
    ]
}